split feature tests up
